+ Changes to reflect fix for MM Bug #4050
Gyro's and Cockpits now have a method in TechConstants to get their TL,
which is based on an int type and an isClan flag.  This required some
parts to be updated.  Also, the Engine techlevel method was updated to
take a year, so Engine parts needed to be updated as well.