Added RAT Table access to GM tools.
From this you can both roll for rat, and if GM mode is enabled, add
directly into your campaign hanger.

Additionally, GMTools is no longer held in memory when not shown and
some utility functions have been added to the UnitTableData for going
from string to int for the weight, unit, and quality.