Fix for Issue #113 - MekHQ 0.3.29 Dates are cut off in Personnel log display
This is a fix for Issue #113.  This only seemed to be affecting Mac OSX. The default font, Lucida Grande in this case, seems to render wider on OSX than it does on other OS's. The solution is to widen the column slightly to accomodate the wider font.  Will need to test that this doesn't have an adverse effect on other OS's.