Mass Repair/Salvage
Mass Repair/Salvage

I created a way to do configurable mass repair/salvage. Having used
MekHQ for over 200 campaign fights, the time to perform repairs is to
most time consuming part. Looking to improve that, I've added this small
feature.

In essense, it creates adds a menu item when right clicking on a unit in
the repair tab. The menu item allows to perform all possible (as
configured) repairs or salvages.

Within the campaign configuration is additional configurations that
allow a user to determine what things are elligible for mass action. For
each of the groups (see below for groups), you can configure if it's
active, what the min skill level to handle it is and the min/max BTHs.

The groups are:
- Armor
- Ammo
- Weapons
- Mek/Tank Locations
- Gyros
- Engines
- Actuators
- Electronics (Cockpit, life support, sensors)
- Other

Additionally, you can configure if the system will:
- use extra time lower the BTH for a tech to the desired target
- use 'rush job' to increase the BTH on easy jobs (i.e., an elite
repairing armor)
- allow carry-over to the next day(s), i.e., allowing a tech to perform
a task that will tke more than their currently available tech minutes

The system, when activated, will find all parts needing repair/salvage
and all techs that could work on it based upon skill type. It then
determines, for each tech, if they can perform the task based upon
configured minumim skill level and target BTHs. If allowed, the system
will attempt to alter the BTH by using extra time/rush job. Once the
list of valid techs is found, the system will pick the one with the
least skill (to aid in increasing techs skills).

I can supply screen shots, etc...