Revert "Finish implementation of the clan_tech_knowledge SPA, and fix all known bugs with the changes to SPAs"
This reverts commit b2c0fb675ad9eaab2600232323f60a4ec00778b1.

Reverting until the feature can be finished, as it breaks what had been
working previously.

New SPAs purchased are stored in the `Person.spas` hash but other than
display in `PersonViewPanel` and the implementation of the Clan Tech
Knowledge SPA, is not used. Loading and saving the campaign and
reinitializing a `Crew` only use the old `options` field.

Fixes #395: SPAs do not appear in Pilot Info in MHQ.