Improved handling of parts in use
The problem is that parts with the same name are all, incorrectly,
lumped into one item on the Parts In Use tab. This means parts like
gyros, which have different tonnages, are listed as which ever tonnage
comes up last. If you were to buy one of these parts, you would only get
the one which is in the map. This change cleans that up to properly
display variable tonnage parts of the same type.

- [Campaign.java] Don't store BA parts
- [PartsStore.java] Update storage of 'by name' map. Previously this
didn't not handle parts with duplicate names very well. Instead of
storing a single part for a given name, it stores all parts in a list
keyed off the name.
- [Part.java] Add method for generating the long description. We
generate names in various ways in multiple places which will hopefully
all eventually be replaced by this long description.
- [PartInUse.java] Remove all code for generating the description and
fall back to new method in Part.java