Better management of rush jobs in MRMS
- Rush jobs apply a reduction on the experience on the tech. This
results in the MRMS system rushing jobs with low BTHs but then making it
too hard for the tech to do. The system now better handles when it moves
a job to a 'rush' mode.
- Fix bug in MRW where it was building the list of available MR types
incorrectly and therefore not applying configuration properly but also
not allow many parts to be fixed via MRW
- Fix newly introduced bug that resets the 'mode' of a part to 'normal'
before calculating which techs can handle it.