Changed all Part classes to use tech progression from MM.
DropshipDockingCollar now has a boolean field to distinguish collars
with no KF boom in anticipation of primitive Dropships, but the value is
not changed.

We have not been distinguishing between IS and Clan endo-steel. I added
a field to distinguish and the PartsStore now stocks both parts, but we
will need campaign post-load processing to set existing parts.