Added implementation of ITechnology interface to Unit.
Use ITechnology interface to filter units by tech level in
UnitSelectorDialog and calculate availability for UnitOrder.

Faction-specific tech progression for units is calculated asynchronously
on load and when a change in campaign options requires calculating a new
faction. Data calculated for previous factions is retained, and can be
used for multiple factions simultaneously.