Fix two tech team-related bugs in Campaign Ops reputation
1. The technicians count was never reset. updatePersonnelCounts()
adds to that variable, and is called when advancing the day and
(possibly) when making other changes, so the number of astechs
required would increase every day.
2. In getSupportDetails(), the code used getAstechPool() (which
only gets the number of temp astechs) instead of getNumberAstechs()
(which takes into account personnel with the role 'astech').
(Only the display was wrong; the calculation used to actually
determine the rating was correct.)