Use dialog for large craft ammo swaps.
Dialog uses MM's BayMunitionsChoicePanel.
Instead of having a designated bin to move capacity to when ammo is
removed as part of an ammo swap, set capacity of all bins according to
what the dialog sets on the Entity, allow negative values for
shotsNeeded, and check other ammo bins in the same bay to determine
there is sufficient capacity before allowing a loading tech action.