Set armor type correctly on armor removed from unit during refit.
The spare part was being created with an uninitialized AT value of -1,
but now uses the AT of location 1 (which is the first location which is
legal on all units).

Note that this does not deal correctly with patchwork armor, but
previously it did not deal correctly with any and now most units will
add the correct armor to the warehouse.

Fixes SF 981:  Armor removed in a refit is not the same as armor needed
to refit back to original