Don't require wait for refit kit if the only needed parts are ammo bins.
Any needed ammo bins are created when the refit begins, and if there is
nothing else on the shopping list the refit can begin immediately. If
there is not enough ammo to fill them, the ammo is added to the shopping
list but does not delay the start of the refit.

SF bug 957 Ammo bin only refit kits