Don't remove new unit parts when a refit is canceled.
This was removing any old unit parts that were unchanged from the
campaign, which was causing an NPE if attempting to refit again, and
even if the new unit part was not part of the old unit it should not be
removed from the warehouse as it has already been purchased, but instead
consolidated with the non-reserved parts.

Any parts or refit kits on the shopping list are now removed if they
have been ordered for a refit that is canceled. Parts in transit are
left alone since they have already been purchased and are on their way.