Fix for issue #626.
This fixes an issue where news items that only list a year for the date would randomly generate the same date for each news item. It was doing this because the NewsItem.finailizeDate(long) method was creating a new Random object with the same seed every time it was called. I just replaced this with megamek.common.Compute.randomInt(int) instead.