MekLab structure tab refreshing incorrectly.
The call to refreshSummary() is supposed to refresh just the weight/slot
breakdown view on the structure tab, but it was calling refresh()
instead. This was causing a listener to get added a second time, so when
the listener was removed to refresh data without updating the Entity it
was updating anyway. This was causing any critical slots to assigned to
internal structure to be unallocated.