Calculate missing unit tech progression on the fly.
Checking the tech progression of a model that did not exist when the
campaign was loaded (e.g. a new configuration in the MekLab) was
returning null where null values should not be allowed. Extracted the
code that performs the calculation and added a check that performs the
calculation and adds to the map if it's missing.