Added null test to check for legal tech level for refits.
A null should not occur unless there is an exception thrown when
calculating the tech level (which would show up in other places as
well), but if that should happen we log it and allow things to proceed
otherwise normally. The exception that led to the null value will be
logged earlier in the file.