Separated side effects from Part#fix()
Restore unit calls fix() on every part whether or not it needs fixing,
presumably for the side effects of resetting other values.
AeroHeatSink#fix() adds a heat sink without checking whether it was
damaged.

I separated the side effects of fix() so they can be called
independently and added a check to AeroHeatSink so it will only add a
heat sink if was actually fixed.