Scale Unit sell values by their Entity's price multiplier
Fix #1497

When computing construction costs, the final cost of many units is scaled by a multiplier.
Generally this is related to the unit's weight class or type (e.g. aerodyne dropships cost 36 times
the sum of their components costs).

This change causes getSellValue() to take this into consideration, generally resulting in higher
sell prices. Ideally the sell value of a perfect condition, undamaged Unit would match it's buy cost.
This change doesn't get all the way there (things like SI are still not accounted for), but this is
a step towards that goal that doesn't require a major redesign of the Entity cost calculation methods.