Improved Awards
This list takes the Default Set, expands the list, and sorts it in order of precedence as they should appear.  The set is called 'standard' to provide better English when awarding them (i.e. 'Awarded Good Conduct of the standard award set' instead of 'Awarded Good Conduct of the Default Set award set').

This adds 'standard.xml' to ...\MekHQ\data\universe\awards\ and the 'standard' folder to ...\MekHQ\data\images\awards\.

I recommend deleting ...\MekHQ\data\universe\awards\Default Set.xml to remove it from the list of award sets, since it is all duplicated and expanded in the standard set.

The ...\MekHQ\data\images\awards\standard folder also includes a ReadMe.txt with this info as well as a description of all included awards, plus an Excel file titled AwardTracker.xlsx with a PowerPivot Table that can be used to import saved .cpnx campaign files to see what awards have been earned for individuals, lances, companies, and battalions based on kills, as long as your TO&E is arranged by battalions, companies, & lances.