Create story arc loading dialog
I put this at the bottom of the check for new campaigns in teh DataLoadingDialog. Ultimately, we want more flexibility here and a story arc should supercede other stuff so probably get its own button in the startup GUI. For now those, this will work to load the story arc into a new campaign and let me test other features.

The dialog basically copies everything from the Campaign Preset Dialog. Eventually, I want to allow it to associate an image with the campaign, but this is good for now.