Track scenarios in StoryEvent and refactor AddScenario.java
I changed AddScenario.java to ScenarioStoryEvent and now I keep track of the scenario inside of the StoryEvent itself which will ultimately make it easier to work with. The scenario also gets assigned the StoryEvent so that they can talk to one another. Note that I will need an alternate way of loading the scenario on save/load once the user activates the scenario or I will end up with two separate scenarios. The sensible way to do this is to track a scenario and mission id in ScenarioStartEvent when it is assigned and then use that instead of generating a new instance when it has a valid number.