Add StoryTrigger and CompleteStoryEventTrigger and associated features
StoryTriggers execute some other thing upon completion of a StoryEvent. A list of default StoryTriggers can be set on a StoryEvent and can be overwritten by a list of StoryTriggers from a StoryOutcome. The only StoryTrigger currently is CompleteStoryEventTrigger which will complete a different StoryEvent. This is the way that MissionEvents can be completed.