Use ScenarioStatus.name() for getResult in ScenarioStoryEvent
This change allows me to use the full range of victory/defeat statuses. To handle the problem of forcing users to create a StoryOutcome for each one, the getResult function now searches for the highest existing ScenarioStatus in the StoryOutcomes hash that is not higher than the actual outcome. This will allow people to set "cut points" to define different outcomes. For example if I use "PYRRHIC_VICTORY" as the only StoryOutcome result, then any victory result would trigger it while draws and defeats would get the default condition. A simple way to get just Victory/Draw/Defeat is to set up StoryOutcomes for PYRRHIC_VICTORY/DRAW/DECISIVE_DEFEAT. Alternatively, the last one can be left out and all defeats will just get the default stuff, if it exists.