Add splash image to StoryDialog
At the moment, I am just loading the old mekhq splash screen here for testing purposes. Eventually, I will want to give the creator more control and the ability to include their own images.

I am not resizing the image, but rather adjusting the size of the dialog to fit the image. So creators will be responsible for creating reasonably sized images. The 500x350 size of the splash images works pretty well as does the 256x350 size of the new default portrait image. So something in that ballpark should generally work well.