Add DateReachedStoryEvent.java
This is a new kind of StoryEvent which is not started by another event but is rather triggered by one of the MMEvents. It will be started (and completed) when the given date is reached in the campaign. It does nothing explicitly although StoryStriggers could be set. It is a transitory even that will most likely be linked to a narrative next event.