Add BotForceRandomizer
This object allows for a simpler method of random generation of bot forces for scenarios. It can be added to any BotForce object and is tracked via XML I/O. It includes variables that parameterize the random generation process to find opfor. These additional random forces are generated and added to the game in GameThread. Currently, I just have a non-interesting test case in the generateForce method, but it should be straightforward to build in the more complex logic needed in subsequent commits.