Added adjusted weight balancing to BotForceRandomizer
Instead of BV, this technique tries to balance the weight of units after applying some modifiers based on unit type. So tank weight only counts for 60% of mech weight and so on. These multipliers are currently hard-coded into the static getSimplePointScore method, but could be made customizable at some point.