Add variation in weight class to BotForceRandomizer
I also added the apache common math library to build.gradle so that I can use their GammaDistribution function to create variation. The center of this distribution is either a focal weight class provided by the object or the mean weight class of the player units. The gamma distribution then samples around this value and the results are trimmed to plausible values and then rounded to an integer.