Create StoryArcStub for selecting and loading story arcs
I was previously just loading all of the possible story arcs for the selector. However, because campaign and version were null at this point, I was passing on both of these null objects into other loaders which was super dangerous and eventually led to problems with loading BotForces in scenarios. So now I load a StoryArcStub instead which just has basic descriptive information and use the selected stub to load the actual story arc once it is selected and after a campaign has been loaded. I also added a version attribute to the top level of the StoryArc xml file to deal with Version.