Add SetDateStoryTrigger class
This StoryTrigger will change the date on a DateReachedStoryPoint. It can be changed with an alternative static date or by the number of days into the future from whenever  the StoryTrigger is executed. This can be used for example when an event should cause another event to happen the next day or in a week's time.