Add CheckMoreScenariosStoryPoint class
This class can be used to check if more active scenarios exist in a given mission. It can be used when creator gives out a set of scenarios in a mission but does not know in what order they will be completed. Along with this, I also allow for the CompleteMission trigger to not send a mission status in which case sucess or failure is determined by the percentWin double (by default 0.5) of the MissionStoryPoint and the percent of actual scenarios that were successful.