This PR fixes the Bad Event StratCon modifier to add a working bonus.  It did not appear to do anything before based on my observations and other bug report.  If you have a "Bad Event" then you should also have another negative modifier.  That didnt happen.
I changed bad event to add +1 equipment quality and 1 hostile BV budget increase.  So it is now a combination of Good Equipment and Hostile BV budget increase.

Ill be honest - I cannot tell if the BV budget increase is working.  I copied the other modifier exactly, but I ran many test with spreadsheet tracking and changed the number to extremes and I could see no patter whatsoever.  But if it does work, then its part of this modifier.  If it doesnt, then Good Equipment will still provide a bonus.