Added Clan education and Prep School functionality
Enhanced the ServiceLogger with methods to log Clan-based education events such as The Blooding and adding specific cases for Trial of Position failures. Also added boolean flags to differentiate institutions as Clan academy or a Prep School. Next, expanded the Academy constructor to include flags for Clan and Prep School. Updated related logic in the EducationController to support Clan education and Prep Schools.