Refactored methods in EducationController and update name generation logic
Modified the getSimplifiedTravelTime function to be non-static and use the current campaign object instead of passing it as an argument. Refactored the getAcademy method in EducationController to accept two parameters - academySetName and academyNameInSet. Enhancements were made to the generateName method to account for local and Clan academy types. Also, fixed values in Education.properties.