Updated academy eligibility conditions and curriculum
Updated academy application conditions in the MekHQ education system, checking whether the campaign's faction has access to Sibkos, and if the applicant is within the right age bracket and is qualified. Also, changed the curriculum of the warrant officer academy from 'Leadership, Negotiation' to 'Leadership, Administration' and lowered the education level requirement for the officer academy.