Updated education module and campaign options
Various updates and fixes have been applied to the education module. The PersonViewPanel now contains more detailed information related to the education of a character. This includes the character's education duration, and their travel to and from the academy. Changes to the campaign options include adjustments to the default values of some settings.