Rolled back awards image picker in PersonViewPanel
This commit refactors how MHQ calculates the number of awards and retrieves the relevant files for medals, ribbons, and misc awards. There is no longer a need for a separate method to process the filename string. The number of awards is now divided by the size of the award tier to create a new variable, awardTierCount, which directly retrieves the correct filename.