Added administrative strain tracking and removed leadership modifiers
Implemented the addition of administrative strain tracking as part of the retirement and defection tracking system. This allows limiting the number of personnel that can be in the unit before incurring a turnover penalty based on admin/HR or admin/Command administration skill. At the same time, removed the usage of leadership modifiers.