Refactored theft functionality and added petty theft
Refactoring was done in the campaign options to divide theft into two categories: theft of units and theft of money. An integer value for average heist percentage was also added, along with the corresponding user interface elements. The handling of petty theft was added with a comprehensive list of possible stolen items.