Added mutiny handling mechanics and related dialog
A mechanic to handle mutiny scenarios among personnel in the MekHQ campaign has been implemented. This includes the addition of a new dialog box providing the user with the options to either support the loyalists or the rebels during such an event. Updated the method to process desertions and added operation for civil war loyalty checks. The Morale.properties file is also modified to include appropriate text messages for the mutiny situations.