Add morale manipulation tools to GMToolsDialog
A General Tools Panel is added to the GMToolsDialog for manipulating morale values. Morale is now adjustable via a JSpinner. This change includes the addition of a tooltip that informs users that changes may not be reflected until they Advance Day or reload the client. The Campaign.java class is also updated to handle morale as a double instead of Double.