Added homeschooling support to education system
The education journey was modified to accommodate the possibility of a person being homeschooled as part of the education system. This includes skipping the journey to/from the academy, adjusting graduation roll calculations, and modifying reports to indicate homeschooling where relevant. Homeschool behavior is determined by a new 'isHomeSchool' academy attribute.