Update static faction references with appropriate isX() call
In order to facilitate custom mercenary factions, update the
mission and contract generation to use the isMercenary() function.

While I was making the changes, I also changed the "PIR" and "REB"
references with isPirate and isRebel so as to avoid having mixed
call styles.

Also added required mock functions to the unit tests.

resolves MegaMek/mekhq#4318