Added personality generation to various personnel creators
The PersonalityController.generatePersonality method was invoked in various personnel creators such as the AbstractCompanyGenerator, HireBulkPersonnelDialog, Campaign, and AbstractProcreation classes. This introduction allows for the generation of personalities for personnel throughout different aspects of the application. Furthermore, changes were made for the biographical descriptions, making use of gender pronouns instead of using the first name repetitively.