Updated character ambition traits in MekHQ
Extended the character ambition traits in the MekHQ game. The previous list contained 7 traits but the updated version now includes 30 different ambition traits. These traits are added to the enum 'Ambition' along with their corresponding descriptions updated in mekhq/resources/Personnel.properties file. Also added for each trait are methods to parse to and from string, and to and from integer.