Added re-enrollment functionality to education system
The education system in the ServiceLogger and EducationController has been expanded to support re-enrollment of a person into a campaign. This includes updating their education information and logging details of the re-enrollment. A new function was also added in the PersonnelTableMouseAdapter to handle commands related to re-enrollment. This overall change enhances the feature set of the education system, giving more flexibility in managing personnel's education journey.