Add command to override force formation level
Added a new command `COMMAND_OVERRIDE_FORCE_FORMATION_LEVEL` to allow the override of the formation level for forces. Implemented `isAddFormationLevel` helper method to check if a formation level can be added to a menu based on the faction. Updated the popup menu to include an option for overriding formation level for eligible forces.