Refactored Years of Service Calculation for autoAwards Time Awards
Updated the `getYearsInService` method in the `Person` class to return a long instead of integer. This ensures accurate representation of service years even for extended durations and refactored related code in `TimeAwards` to take advantage of this method, as it didn't exist when TimeAwards were originally written.Add method to calculate years in service for a character

Introduced getYearsInService method in Person.java to determine the time a character has been employed in a campaign. This method accounts for the recruitment date and factors in the date of death or retirement to