Added Contract Difficulty Estimate, Including GUI Support
Refactored the `AtBContract` class to include methods for calculating and displaying contract difficulty. Added UI elements in `ContractMarketDialog` and `ContractSummaryPanel` to show difficulty estimates.

Added dialog calls for when the player is picking unusually easy or overly challenging contracts.

Enhanced code maintainability by removing a long switch statement for faction logos and replaced it with a function call. Expanded the switch to include all factions we have logos for, and not just the Clan factions.