Change refit dialog to refit/customize dialog that allows style of refit
Using a refit kit is easier than and has different rules from customizing
a mek, and until now the only way to use the customization rules was to use
the meklab manually.

This setup also allows you to customize a mek to a non-canon variant but
only refit to a canon variant if canon-refits-only is turned on. This means
that when that option is enabled it is much easier to repeat saved meklab
customizations.

This further means that building mek variants in MegaMekLab and then carrying
out the customization as if you had done it inside MekHQMekLab is possible.

Utilities.getAllVariants was changed but that function only has the code
we're working on here calling it.