Prevent Early Clan Units in Non-Clan Campaigns
Added a check to prevent Clan units from appearing in the unit market for non-Clan campaigns, prior to the fifth wave of invasion.

This ensures historical accuracy by removing early access to Clan technology.

I chose the fifth wave of invasion, as that seems a reasonable point by which the proliferation of Clan technology (while not common) would likely begin across the Inner Sphere.