Disabled Commit Forces Button when No Forces Selected
In StratCon selecting a scenario with no forces assigned allows you to select a primary force, for that scenario, however if no forces are selected and you press 'commit' the scenario's state is changed as if a force had been committed - meaning all future forces are considered Reinforcements. This change disables the commit button if no forces have been selected. This dialog can be exited if users don't want to commit a force, by clicking the exit window button.